Passing Test: Added test to verify type and number of coins returned.
The previous test didn't account for a customer entering multiple quarters and getting back the
same number of quarters. It just checked that a quarter was returned. This test ensures the customer
will get back the same number and type of coins they entered. Short changing the customer would be bad :)